+++ packages/cubejs-playground/src/QueryBuilder/FilterInput.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import * as PropTypes from 'prop-types';
 import { Input } from 'antd';
 import { Select } from '../components';