+++ packages/cubejs-playground/src/components/Pivot/Item.js
@@ -1,4 +1,3 @@
-import React from 'react';
 import { Draggable } from 'react-beautiful-dnd';
 import { DragOutlined } from '@ant-design/icons';
 import { Typography } from 'antd';