+++ packages/cubejs-playground/src/index.js
@@ -1,14 +1,12 @@
-import React from 'react';
 import ReactDOM from 'react-dom';
 import { Router, Route } from 'react-router-dom';
 import { createHashHistory } from 'history';
 import IndexPage from './IndexPage';
-import ExplorePage from './ExplorePage';
 import SchemaPage from './SchemaPage';
-import DashboardPage from './DashboardPage';
 import App from './App';
 import { page } from './events';
 import TemplateGalleryPage from './TemplateGallery/TemplateGalleryPage';
+import { ExplorePage, DashboardPage, ConnectionWizardPage } from './pages';
 
 const history = createHashHistory();
 history.listen((location) => {
@@ -21,6 +19,11 @@ ReactDOM.render(
       <Route key="index" exact path="/" component={IndexPage} />
       <Route key="build" path="/build" component={ExplorePage} />
       <Route key="schema" path="/schema" component={SchemaPage} />
+      <Route
+        key="connection"
+        path="/connection"
+        component={() => <ConnectionWizardPage history={history} />}
+      />
       <Route key="dashboard" path="/dashboard" component={DashboardPage} />
       <Route
         key="template-gallery"