+++ packages/cubejs-playground/src/pages/Explore/ExplorePage.js
@@ -1,11 +1,11 @@
-/* global window */
-import React, { Component } from 'react';
+import { Component } from 'react';
 import cubejs from '@cubejs-client/core';
 import { CubeProvider } from '@cubejs-client/react';
 import { fetch } from 'whatwg-fetch';
 import PropTypes from 'prop-types';
-import DashboardSource from './DashboardSource';
-import PlaygroundQueryBuilder from './PlaygroundQueryBuilder';
+
+import DashboardSource from '../../DashboardSource';
+import PlaygroundQueryBuilder from '../../PlaygroundQueryBuilder';
 
 class ExplorePage extends Component {
   constructor(props) {