+++ packages/cubejs-server-core/src/index.ts
@@ -1,5 +1,6 @@
 import { CubejsServerCore } from './core/server';
 
+export * from './core/logger';
 export * from './core/server';
 export * from './core/types';
 export * from './core/FileRepository';