+++ packages/cubejs-server/src/server/container.ts
@@ -289,15 +289,29 @@ export class ServerContainer {
       // eslint-disable-next-line no-loop-func
       process.on(bindSignal, async (signal) => {
         process.exit(
-          await server.shutdown(signal)
+          await server.shutdown(signal, true)
         );
       });
     }
 
+    let restartHandler: Promise<0|1>|null = null;
+
     process.addListener('SIGUSR1', async (signal) => {
       console.log(`Received ${signal} signal, reloading`);
 
-      await server.shutdown(signal, true);
+      if (restartHandler) {
+        console.log(`Unable to restart server while it's already restarting`);
+
+        return;
+      }
+
+      try {
+        restartHandler = server.shutdown(signal, true);
+
+        await restartHandler;
+      } finally {
+        restartHandler = null;
+      }
 
       server = await makeInstance();
     });